refactor(scripts): file-organization sweep over the services, models and workspace roots
Eight suites computed the repository root from their own depth, and every one of them now reads REPOSITORY_ROOT, so a suite that moves a directory deeper no longer resolves a root that is not there. runGit had one consumer and moves under services/coderabbit beside it. The two repo-invariant suites that sat at the source root — the agent directory exclusions and the ledger scopes — join the other invariants in workspace/, and their pure helpers hoist to module scope now that they capture nothing; the skills and docs page citing them follow.
