+++ scripts/src/services/coderabbit/window/getFileCount.ts
@@ -1,4 +1,4 @@
+import { runGit } from "#src/services/coderabbit/runGit";
 import { getNonEmptyLines } from "#src/services/getNonEmptyLines";
-import { runGit } from "#src/services/runGit";
 
 export const getFileCount = (range: string): number => getNonEmptyLines(runGit(["diff", "--name-only", range])).length;
