+++ scripts/src/workspace/declarationGeneration.test.ts
@@ -1,3 +1,4 @@
+import { REPOSITORY_ROOT } from "#src/services/constants";
 import { getWorkspacePackageDirectories } from "#src/services/getWorkspacePackageDirectories";
 import { parseMachineJson } from "#src/services/parseMachineJson";
 import { existsSync, readFileSync } from "node:fs";
@@ -14,13 +15,12 @@ import { describe, expect, test } from "vitest";
 const readJsonFile = (path: string) => parseMachineJson<Record<string, unknown>>(readFileSync(path, "utf8"));
 
 describe("declaration generation", () => {
-  const repositoryRoot = resolve(import.meta.dirname, "../../..");
   // A package builds with tsdown exactly when it has a tsdown config, so the set is discovered rather than listed:
   // A listed set silently stops covering the package added after it was written, which is the only way this
   // Invariant can be broken. Every member the workspace declares, because two of the ones that build sit under
   // `apps`. (`apps/web` is a Nuxt application, has no tsdown config, and emits nothing.)
-  const PACKAGE_PATHS = getWorkspacePackageDirectories(repositoryRoot).filter((packagePath) =>
-    existsSync(resolve(repositoryRoot, packagePath, "tsdown.config.ts")),
+  const PACKAGE_PATHS = getWorkspacePackageDirectories(REPOSITORY_ROOT).filter((packagePath) =>
+    existsSync(resolve(REPOSITORY_ROOT, packagePath, "tsdown.config.ts")),
   );
   // The one package the invariant cannot cover: an SFC's types cannot be written out by hand, so its declarations
   // Go through vue-tsc by way of `dts.vue` whatever `isolatedDeclarations` says. Any other package joining it is
@@ -54,7 +54,7 @@ describe("declaration generation", () => {
     // Named rather than counted: a package that starts emitting declarations without `isolatedDeclarations` names
     // Itself in the failure. An empty discovery cannot pass this either, since the exception must be present.
     const slowPathPackagePaths = PACKAGE_PATHS.filter((packagePath) => {
-      const packageDirectory = resolve(repositoryRoot, packagePath);
+      const packageDirectory = resolve(REPOSITORY_ROOT, packagePath);
       return (
         readEmitsDeclarations(packageDirectory) && !readIsolatedDeclarations(resolve(packageDirectory, "tsconfig.json"))
       );
